+++ apps/documentation/src/app/stories/main/tree.stories.ts
@@ -2,21 +2,26 @@ import { Meta, Story, moduleMetadata } from '@storybook/angular';
 import { Ui5WebcomponentsModule, TreeComponent } from '@ui5/webcomponents-ngx';
 import { extractArgTypes } from '../../arg-type-tools';
 
-const description = `### Overview The <code>ui5-tree</code> component provides a tree structure for displaying data in a hierarchy.
+const description = `
+### Overview
 
-<h3>Usage</h3>
+The <code>ui5-tree</code> component provides a tree structure for displaying data in a hierarchy.
+
+### Usage
 
 <h4>When to use:</h4> <ul> <li>To display hierarchically structured items.</li> <li>To select one or more items out of a set of hierarchically structured items.</li> </ul>
 
 <h4>When not to use:</h4> <ul> <li>To display items not hierarchically strcutured. In this case, use the List component.</li> <li>To select one item from a very small number of non-hierarchical items. Select or ComboBox might be more appropriate.</li> <li>The hierarchy turns out to have only two levels. In this case, use List with group items.</li> </ul>
 
-<h3>Keyboard Handling</h3>
+### Keyboard Handling
 
 The <code>ui5-tree</code> provides advanced keyboard handling. The user can use the following keyboard shortcuts in order to navigate trough the tree: <ul> <li>[UP/DOWN] - Navigates up and down the tree items that are currently visible.</li> <li>[RIGHT] - Drills down the tree by expanding the tree nodes.</li> <li>[LEFT] - Goes up the tree and collapses the tree nodes.</li> </ul> <br>
 
 The user can use the following keyboard shortcuts to perform selection, when the <code>mode</code> property is in use: <ul> <li>[SPACE] - Selects the currently focused item upon keyup.</li> <li>[ENTER] - Selects the currently focused item upon keydown.</li> </ul>
 
-<h3>ES6 Module Import</h3> <code>import { TreeComponent } from "@ui5/webcomponents-ngx/main/tree";</code>`;
+### ES6 Module Import
+
+<code>import { TreeComponent } from "@ui5/webcomponents-ngx/main/tree";</code>`;
 export default {
   title: 'UI5 Web Components / Main / Tree',
   component: TreeComponent,