+++ libs/ui5-angular/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.3.0-rc.10](https://github.com/SAP/ui5-webcomponents-ngx/compare/v0.3.0-rc.9...v0.3.0-rc.10) (2024-07-09)
+
+**Note:** Version bump only for package @ui5/webcomponents-ngx
+
+# [0.3.0-rc.9](https://github.com/SAP/ui5-webcomponents-ngx/compare/v0.3.0-rc.8...v0.3.0-rc.9) (2024-06-25)
+
+**Note:** Version bump only for package @ui5/webcomponents-ngx
+
+# [0.3.0-rc.8](https://github.com/SAP/ui5-webcomponents-ngx/compare/v0.3.0-rc.7...v0.3.0-rc.8) (2024-06-20)
+
+**Note:** Version bump only for package @ui5/webcomponents-ngx
+
 # [0.3.0-rc.7](https://github.com/SAP/ui5-webcomponents-ngx/compare/v0.3.0-rc.6...v0.3.0-rc.7) (2024-06-20)
 
 ### Features